% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{arrow_read_dataset}
\alias{arrow_read_dataset}
\title{Safely use arrow to read a Parquet file}
\usage{
arrow_read_dataset(filename)
}
\arguments{
\item{filename}{A local Parquet file}
}
\value{
An \code{arrow::Dataset}
}
\description{
This function handles some failure modes, including if the Parquet file is
corrupted.
}
\keyword{internal}
