% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_gpkg}
\alias{download_gpkg}
\title{Download geopackage to tempdir}
\usage{
download_gpkg(filesD)
}
\arguments{
\item{filesD}{A string with the url address of a geobr dataset}
}
\description{
Download geopackage to tempdir
}
\seealso{
Other support functions: 
\code{\link{select_data_type}()}
}
\concept{support functions}
