% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_metadata.R
\name{download_metadata}
\alias{download_metadata}
\title{Download data.framte with a summary of the data sets included in grobr}
\usage{
download_metadata()
}
\description{
Download data.framte with a summary of the data sets included in grobr
}
\examples{
\donttest{

library(geobr)

df <- download_metadata()

}

}
\concept{general support functions}
