% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_arrw}
\alias{filter_arrw}
\title{Filter data set to return specific states}
\usage{
filter_arrw(temp_arrw = parent.frame()$temp_arrw, code)
}
\arguments{
\item{temp_arrw}{An internal arrow table}

\item{code}{The two-digit code of a state or a two-letter uppercase
abbreviation (e.g. 33 or "RJ"). If \code{code_state="all"} (the
default), the function downloads all states.}
}
\value{
A simple feature \code{sf} or \code{data.frame}.
}
\description{
Filter data set to return specific states
}
\keyword{internal}
