% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_state}
\alias{filter_state}
\title{Filter data set to return specific states}
\usage{
filter_state(
  temp_sf = parent.frame()$temp_sf,
  code = parent.frame()$code_state
)
}
\arguments{
\item{temp_sf}{An internal simple feature or data.frame}

\item{code}{The two-digit code of a state or a two-letter uppercase
abbreviation (e.g. 33 or "RJ"). If \code{code_state="all"} (the
default), the function downloads all states.}
}
\value{
A simple feature \code{sf} or \code{data.frame}.
}
\description{
Filter data set to return specific states
}
\keyword{internal}
