% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geobr_open_dataset}
\alias{geobr_open_dataset}
\title{Safely opens a Parquet file}
\usage{
geobr_open_dataset(filename)
}
\arguments{
\item{filename}{A local Parquet file}
}
\value{
An \code{duckspatial_df}
}
\description{
This function handles some failure modes, including if the Parquet file is
corrupted.
}
\keyword{internal}
