% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_online}
\alias{is_online}
\title{Check internet connection with Ipea server}
\usage{
is_online(
  file_url = "https://www.ipea.gov.br/geobr/metadata/metadata_gpkg.csv"
)
}
\arguments{
\item{file_url}{A string with the file_url address of an aop dataset}
}
\value{
Logic \verb{TRUE or }FALSE`.
}
\description{
Checks if there is internet connection to Ipea server to download aop data.
}
\seealso{
Other support functions: 
\code{\link{download_gpkg}()},
\code{\link{load_gpkg}()},
\code{\link{select_data_type}()},
\code{\link{select_metadata}()},
\code{\link{select_year_input}()}
}
\concept{support functions}
