% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_muni.R
\name{lookup_muni}
\alias{lookup_muni}
\title{Look up municipality codes and names}
\usage{
lookup_muni(year, name_muni = NULL, code_muni = NULL)
}
\arguments{
\item{year}{Numeric. Year of the data in YYYY format. It defaults to \code{NULL}
and reads the data from the latest year available.}

\item{name_muni}{The municipality name to be looked up.}

\item{code_muni}{The municipality code to be looked up.}
}
\value{
A \code{data.frame} with 13 columns identifying the geographies information
of that municipality.

A \code{data.frame}
}
\description{
Input a municipality \strong{name} \emph{or} \strong{code} and get the names
and codes of the municipality.
}
\details{
Only available from 2010 Census data so far
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) withAutoprint(\{ # examplesIf}
# Look for municipality Rio de Janeiro
mun <- lookup_muni(
  name_muni = "Rio de Janeiro",
  year = 2022
  )

# Look for a given municipality code
mun <- lookup_muni(
  code_muni = 3304557,
  year = 2022
  )

# Get the lookup table for all municipalities
mun_all <- lookup_muni(
  name_muni = "all",
  year = 2022
  )

# Or:
mun_all <- lookup_muni(
  code_muni = "all",
  year = 2022
  )
\dontshow{\}) # examplesIf}
}
