% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numbers_only}
\alias{numbers_only}
\title{Check if vector only has numeric characters}
\usage{
numbers_only(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
Logical. \code{TRUE} if vector only has numeric characters.
}
\description{
Checks if vector only has numeric characters
}
\keyword{internal}
