% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_areaponderacao.R
\name{read_areaponderacao}
\alias{read_areaponderacao}
\title{Title}
\usage{
read_areaponderacao(CODE = NULL, year = NULL)
}
\arguments{
\item{year}{}
}
\description{
Title
}
\examples{
\dontrun{
# Exemplos
dados <- read_areaponderacao(year=2010)
dados <- read_areaponderacao(3500000,2010)
dados <- read_areaponderacao(123,2010)
dados <- read_areaponderacao("df",2010)
dados <- read_areaponderacao(1302603,2010)
dados <- read_areaponderacao(35)
dados <- read_areaponderacao(14,2010)
dados <- read_areaponderacao()

# mapa
library(mapview)
mapview(dados)
}




}
