% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_comparable_areas.R
\name{read_comparable_areas}
\alias{read_comparable_areas}
\title{Download spatial data of historically comparable municipalities}
\usage{
read_comparable_areas(
  start_year = 1970,
  end_year = 2010,
  simplified = TRUE,
  showProgress = TRUE,
  cache = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{start_year}{Numeric. Start year to the period in the YYYY format.
Defaults TO \code{1970}.}

\item{end_year}{Numeric. End year to the period in the YYYY format. Defaults
to \code{2010}.}

\item{simplified}{Logic \code{FALSE} or \code{TRUE}, indicating whether the function
should return the data set with 'original' spatial resolution or a data set
with 'simplified' geometry. Defaults to \code{TRUE}. For spatial analysis and
statistics users should set \code{simplified = FALSE}. Borders have been
simplified by removing vertices of borders using \verb{st_simplify\{sf\}} preserving
topology with a \code{dTolerance} of 100.}

\item{showProgress}{Logical. Defaults to \code{TRUE} display progress bar.}

\item{cache}{Logical. Whether the function should read the data cached
locally, which is faster. Defaults to \code{cache = TRUE}. By default,
\code{geobr} stores data files in a temporary directory that exists only
within each R session. If \code{cache = FALSE}, the function will download
the data again and overwrite the local file.}

\item{verbose}{A logical. If \code{TRUE} (the default), the function prints
informative messages and shows download progress bar. If \code{FALSE}, the
function is silent.}
}
\value{
An \verb{"sf" "data.frame"} OR an \code{ArrowObject}
}
\description{
This function downloads the shape file of minimum comparable area of
municipalities, known in Portuguese as 'Areas minimas comparaveis (AMCs)'.
The data is available for any combination of census years between 1872-2010.
These data sets are generated based on the Stata code originally developed by
Ehrl (2017) \doi{10.1590/0101-416147182phe}, and translated into \code{R} by the
\code{geobr} team.
}
\details{
These data sets are generated based on the original Stata code developed by
Philipp Ehrl. If you use these data, please cite:
\itemize{
\item Ehrl, P. (2017). Minimum comparable areas for the period 1872-2010: an
aggregation of Brazilian municipalities. Estudos Econômicos (São Paulo),
47(1), 215-229. https://doi.org/10.1590/0101-416147182phe
}
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) withAutoprint(\{ # examplesIf}
  amc <- read_comparable_areas(start_year=1970, end_year=2010)
\dontshow{\}) # examplesIf}
}
