% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mesoregiao.R
\name{read_mesorregiao}
\alias{read_mesorregiao}
\title{Download shape files of meso region.}
\usage{
read_mesorregiao(year = NULL, cod_meso = NULL)
}
\arguments{
\item{year}{the year of the data download (defaults to 2010)}

\item{cod_meso}{}

\item{cod_meso}{x-digit code of the meso region. If a two-digit code of a state is passed,
the function will load all meso regions of that state. If cod_meso="all", all meso regions of the country are loaded.}
}
\description{
Download shape files of meso region.
}
\examples{
\dontrun{

library(geobr)

# Read all meso regions of a state at a given year
  meso <- read_mesorregiao(cod_meso=12, year=2017)

# Read all meso regions of the country at a given year
  meso <- read_mesorregiao(cod_meso="all", year=2010)
}

}
\seealso{
Other general area functions: \code{\link{read_microregiao}},
  \code{\link{read_municipio}}, \code{\link{read_uf}}
}
