% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_micro_region.R
\name{read_micro_region}
\alias{read_micro_region}
\title{Download spatial data of micro regions}
\usage{
read_micro_region(
  year,
  code_micro = "all",
  simplified = TRUE,
  output = "sf",
  showProgress = TRUE,
  cache = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{year}{Numeric. Year of the data in \code{YYYY} format.}

\item{code_micro}{5-digit code of a micro region. If the two-digit code or a
two-letter uppercase abbreviation of a state is passed, (e.g. 33 or
"RJ") the function will load all micro regions of that state. If
\code{code_micro="all"} (Default), the function downloads all micro regions
of the country.}

\item{simplified}{Logic \code{FALSE} or \code{TRUE}, indicating whether the function
should return the data set with 'original' spatial resolution or a data set
with 'simplified' geometry. Defaults to \code{TRUE}. For spatial analysis and
statistics users should set \code{simplified = FALSE}. Borders have been
simplified by removing vertices of borders using \verb{st_simplify\{sf\}} preserving
topology with a \code{dTolerance} of 100.}

\item{output}{String. Type of object returned by the function. Defaults to
\code{"sf"}, which loads the data into memory as an sf object. Alternatively,
\code{"duckdb"} returns a lazy spatial table backed by DuckDB via the
duckspatial package, and \code{"arrow"} returns an Arrow dataset. Both
\code{"duckdb"} and \code{"arrow"} support out-of-memory processing of large
data sets.}

\item{showProgress}{Logical. Defaults to \code{TRUE} display progress bar.}

\item{cache}{Logical. Whether the function should read the data cached
locally, which is faster. Defaults to \code{cache = TRUE}. By default,
\code{geobr} stores data files in a temporary directory that exists only
within each R session. If \code{cache = FALSE}, the function will download
the data again and overwrite the local file.}

\item{verbose}{A logical. If \code{TRUE} (the default), the function prints
informative messages and shows download progress bar. If \code{FALSE}, the
function is silent.}
}
\value{
An \verb{"sf" "data.frame"} OR an \code{ArrowObject}
}
\description{
Data at scale 1:250,000, using Geodetic reference system "SIRGAS2000" and CRS(4674)
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) withAutoprint(\{ # examplesIf}
# Read an specific micro region a given year
micro <- read_micro_region(code_micro=11008, year=2018)

# Read micro regions of a state at a given year
micro <- read_micro_region(code_micro="AM", year=2018)
micro <- read_micro_region(code_micro=12, year=2018)

# Read all micro regions at a given year
 micro <- read_micro_region(code_micro="all", year=2018)
\dontshow{\}) # examplesIf}
}
