% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_microregiao.R
\name{read_microregiao}
\alias{read_microregiao}
\title{Download shape files of micro region.}
\usage{
read_microregiao(year = NULL, cod_micro = NULL)
}
\arguments{
\item{year}{the year of the data download (defaults to 2010)}

\item{cod_micro}{x-digit code of the micro region. If a two-digit code of a state is passed,
the function will load all micro regions of that state. If cod_micro="all", all micro regions of the country are loaded.}
}
\description{
Download shape files of micro region.
}
\examples{
\dontrun{

library(geobr)

# Read all micro regions of a state at a given year
  micro <- read_microregiao(cod_micro=12, year=2017)

# Read all micro regions at a given year
  micro <- read_microregiao(cod_micro="all", year=2010)
}


}
\seealso{
Other general area functions: \code{\link{read_mesorregiao}},
  \code{\link{read_municipio}}, \code{\link{read_uf}}
}
