% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_municipality_local.R
\name{read_municipality2}
\alias{read_municipality2}
\title{Download shape files of municipalities}
\usage{
read_municipality2(cod_muni, year = NULL)
}
\arguments{
\item{cod_muni}{The 7-digit code of a municipality. If the two-digit code of a state is used,
the function will load all municipalities of that state. If cod_muni="all", all municipalities will be loaded.}

\item{year}{Year of the data (defaults to 2010)}
}
\description{
Download shape files of municipalities
}
\examples{
\dontrun{

library(geobr)

# Read specific municipality at a given year
  mun <- read_municipality(cod_muni=1200179, year=2017)

# Read all municipalities of a state at a given year
  mun <- read_municipality(cod_muni=12, year=2010)

}
}
\seealso{
Other general area functions: \code{\link{read_census_tract}},
  \code{\link{read_meso_region}},
  \code{\link{read_micro_region}},
  \code{\link{read_municipality}},
  \code{\link{read_state}},
  \code{\link{read_statistical_grid}},
  \code{\link{read_weighting_area}}
}
