% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_municipio.R
\name{read_municipio}
\alias{read_municipio}
\title{Download shape files of municipalities}
\usage{
read_municipio(year = NULL, cod_mun = NULL)
}
\arguments{
\item{year}{the year of the data download (defaults to 2010)}

\item{cod_mun}{7-digit code of the municipality. If a the two-digit code of a state is used,
the function will load all munuciplaities of that state. If not informed, all municipalities will be loaded.}
}
\description{
Download shape files of municipalities
}
\examples{
\dontrun{

library(geobr)

# Read specific municipality at a given year
  mun <- read_municipio(cod_mun=1200179, year=2017)

# Read all municipalities of a state at a given year
  mun <- read_municipio(cod_mun=12, year=2010)

}
}
\seealso{
Other general area functions: \code{\link{read_uf}}
}
