% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_setorcensitario.R
\name{read_setorcensitario}
\alias{read_setorcensitario}
\title{Download shape files of census sectors of the Brazilian Population Census}
\usage{
read_setorcensitario(CODE = NULL, year = NULL, zone = "urban")
}
\arguments{
\item{CODE}{One can either pass the 7-digit code of a Municipality or the 2-digit code of a State.If CODE="all", all census sectors of the country are loaded.}

\item{year}{the year of the data download (defaults to 2010)}

\item{zone}{"urban" or "rural" for separation in the year 2000}
}
\description{
Download shape files of census sectors of the Brazilian Population Census
}
\examples{
\dontrun{
# Exemplos
dados <- read_setorcensitario(year=2010)
dados <- read_setorcensitario(123,2010)
dados <- read_setorcensitario("df",2010)
dados <- read_setorcensitario(1302603,2010)
dados <- read_setorcensitario(35)
dados <- read_setorcensitario(14,2010)
dados <- read_setorcensitario()

# mapa
library(mapview)
mapview(dados)
}




}
\seealso{
Other general area functions: \code{\link{read_areaponderacao}},
  \code{\link{read_mesorregiao}},
  \code{\link{read_microregiao}},
  \code{\link{read_municipio}}, \code{\link{read_uf}}
}
\concept{general area functions}
