% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_uf.R
\name{read_uf}
\alias{read_uf}
\title{Download shape files of state}
\usage{
read_uf(year = NULL, cod_uf = NULL)
}
\arguments{
\item{year}{the year of the data download (defaults to 2010)}

\item{cod_uf}{2-digit state code. If not informed, all states will be loaded.}
}
\description{
Download shape files of state
}
\examples{
\dontrun{

library(geobr)

# Read specific municipality at a given year
  mun <- read_municipio(cod_uf=12, year=2017)

# Read al lstate at a given year
  mun <- read_municipio(cod_mun=12, year=2010)

}
}
\seealso{
Other general area functions: \code{\link{read_municipio}}
}
