% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_islands.R
\name{remove_islands}
\alias{remove_islands}
\title{Remove islands from Brazil}
\usage{
remove_islands(x)
}
\arguments{
\item{x}{An 'sf' object with CRS EPSG:4674. Usually an object returned from
other geobr functions, such as \code{read_country()}, \code{read_states()},
\code{read_municipality()}, or similar functions.}
}
\value{
An \code{sf} data frame with the same attributes as \code{x}, but with distant
islands removed from the geometry.
}
\description{
Removes Brazilian islands that are approximately more than 20 km from the
mainland coast. This is useful when analyses or data visualization should
focus on the continental territory of Brazil and exclude distant oceanic
islands.
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) withAutoprint(\{ # examplesIf}
library(geobr)
library(sf)

br <- read_country(year = 1991)

br_no_islands <- remove_islands(br)

plot(br)
\dontshow{\}) # examplesIf}
}
