% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_geometry_type}
\alias{select_geometry_type}
\title{Select data type: 'original' or 'simplified' (default)}
\usage{
select_geometry_type(temp_meta, simplified_geometry)
}
\arguments{
\item{temp_meta}{A data.frame with the metadata of geobr datasets}

\item{simplified_geometry}{Logical \code{TRUE} or \code{FALSE} indicating  whether the
function should return a dataset with the 'original' geometry or a
dataset with 'simplified' geometry (Defaults to \code{TRUE})}
}
\description{
Select data type: 'original' or 'simplified' (default)
}
\keyword{internal}
