% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_connection}
\alias{check_connection}
\title{Check internet connection with Ipea server}
\usage{
check_connection(
  url = "https://www.ipea.gov.br/geobr/metadata/metadata_gpkg.csv",
  silent = FALSE
)
}
\arguments{
\item{url}{A string with the url address of an aop dataset}

\item{silent}{Logical. Throw a message when silent is \code{FALSE} (default)}
}
\value{
Logical. \code{TRUE} if url is working, \code{FALSE} if not.
}
\description{
Checks if there is an internet connection with Ipea server.
}
\keyword{internal}
