% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_connection}
\alias{check_connection}
\title{Check internet connection with Ipea server}
\usage{
check_connection(
  file_url = "https://www.ipea.gov.br/geobr/metadata/metadata_gpkg.csv"
)
}
\arguments{
\item{file_url}{A string with the file_url address of an geobr dataset}
}
\value{
Logical. \code{TRUE} if url is working, \code{FALSE} if not.
}
\description{
Check internet connection with Ipea server
}
\keyword{internal}
