% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_metadata.R
\name{download_metadata}
\alias{download_metadata}
\title{Support function to download metadata internally used in geobr}
\usage{
download_metadata()
}
\description{
Support function to download metadata internally used in geobr
}
\examples{
\donttest{

library(geobr)

df <- download_metadata()

}

}
\seealso{
Other general support functions: 
\code{\link{list_geobr}()}
}
\concept{general support functions}
