% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_metadata2}
\alias{download_metadata2}
\title{Support function to download metadata internally used in geobr}
\usage{
download_metadata2()
}
\description{
Support function to download metadata internally used in geobr
}
\examples{
\dontrun{ if (interactive()) {
df <- download_metadata2()
}
}
}
\keyword{internal}
