% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_parquet}
\alias{download_parquet}
\title{Download parquet to tempdir}
\usage{
download_parquet(
  filename_to_download,
  showProgress = parent.frame()$showProgress,
  cache = parent.frame()$cache
)
}
\arguments{
\item{filename_to_download}{A string with the file name}

\item{showProgress}{Logical. Defaults to \code{TRUE} display progress bar.}

\item{cache}{Logical. Whether the function should read the data cached
locally, which is faster. Defaults to \code{cache = TRUE}. By default,
\code{geobr} stores data files in a temporary directory that exists only
within each R session. If \code{cache = FALSE}, the function will download
the data again and overwrite the local file.}
}
\description{
Download parquet to tempdir
}
\keyword{internal}
