% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geobr.R
\docType{package}
\name{geobr}
\alias{geobr}
\title{\code{geobr} package}
\description{
Easy access to shapefiles of the Brazilian Institute of Geography and Statistics (IBGE) and other official spatial data sets of Brazil
}
\details{
See the README on
\href{https://github.com/ipeaGIT/geobr#readme}{GitHub}
}
