% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_state_correspondence_table.R
\docType{data}
\name{grid_state_correspondence_table}
\alias{grid_state_correspondence_table}
\title{A correspondence table indicating what quadrants of IBGE's statistical grid intersect with each Brazilian state}
\format{
A data frame sf with 139 rows and 3 columns
}
\usage{
data(grid_state_correspondence_table)
}
\description{
Built-in dataset


\itemize{
  \item \code{code_uf}: IBGE code of State (2-digit, numeric)
  \item \code{name_state}: Title-case name of state (character)
  \item \code{code_grid}: Unique code of each quadrant of IBGE's statistical grid
}
}
\details{
correspondence table indicating what quadrants of IBGE's statistical grid intersect with each Brazilian state
}
\note{
Last updated 2019-06-17
}
\keyword{datasets}
