% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_geobr.R
\name{list_geobr}
\alias{list_geobr}
\title{List all data sets available in the geobr package}
\usage{
list_geobr(wide = TRUE)
}
\arguments{
\item{wide}{Whether the the output data frame should come in wide (\code{TRUE}) or
long format (\code{FALSE}).}
}
\value{
A \code{data.frame}
}
\description{
Returns a data frame with all datasets available in the geobr package
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) withAutoprint(\{ # examplesIf}
df <- list_geobr()
\dontshow{\}) # examplesIf}
}
\seealso{
Other support functions: 
\code{\link{lookup_muni}()}
}
\concept{support functions}
