% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_geobr.R
\name{list_geobr}
\alias{list_geobr}
\title{List all data sets available in the geobr package}
\usage{
list_geobr()
}
\description{
Returns a data frame with all datasets available in the geobr package
}
\examples{
\donttest{
df <- list_geobr()
}
}
\seealso{
Other general support functions: 
\code{\link{download_metadata}()}
}
\concept{general support functions}
