% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_gpkg}
\alias{load_gpkg}
\title{Load geopackage from tempdir to global environment}
\usage{
load_gpkg(filesD, temps = NULL)
}
\arguments{
\item{filesD}{A string with the url address of a geobr dataset}

\item{temps}{The address of a gpkg file stored in tempdir. Defaults to NULL}
}
\description{
Load geopackage from tempdir to global environment
}
\seealso{
Other support functions: 
\code{\link{download_gpkg}()},
\code{\link{select_data_type}()}
}
\concept{support functions}
