% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_gpkg}
\alias{load_gpkg}
\title{Load geopackage from tempdir to global environment}
\usage{
load_gpkg(temps = NULL)
}
\arguments{
\item{temps}{The address of a gpkg file stored in tempdir. Defaults to NULL}
}
\description{
Load geopackage from tempdir to global environment
}
\keyword{internal}
