% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_biomes.R
\name{read_biomes}
\alias{read_biomes}
\title{Download official data of Brazilian biomes as an sf object.}
\usage{
read_biomes(year)
}
\arguments{
\item{year}{A date number in YYYY format.}
}
\description{
This data set covers the whole of Brazil and it includes the polygons of all of all biomes present in Brazilian
territory. The last update of the data was 2004 (only year for which the data is currently available). The original
data comes from IBGE and can be found at https://geoftp.ibge.gov.br/informacoes_ambientais/estudos_ambientais/biomas/ .
}
\seealso{
Other general area functions: \code{\link{read_census_tract}},
  \code{\link{read_country}},
  \code{\link{read_meso_region2}},
  \code{\link{read_micro_region}},
  \code{\link{read_municipality}},
  \code{\link{read_region}}, \code{\link{read_state}},
  \code{\link{read_statistical_grid}},
  \code{\link{read_weighting_area}}
}
\concept{general area functions}
