% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_capitals.R
\name{read_capitals}
\alias{read_capitals}
\title{Download data of state capitals}
\usage{
read_capitals(as_sf = TRUE, showProgress = TRUE, cache = TRUE, verbose = TRUE)
}
\arguments{
\item{as_sf}{Logic. If \code{TRUE} (the default), the function returns an
\verb{sf data.frame}. If \code{FALSE}, the function returns an arrow dataset.}

\item{showProgress}{Logical. Defaults to \code{TRUE} display progress bar.}

\item{cache}{Logical. Whether the function should read the data cached
locally, which is faster. Defaults to \code{cache = TRUE}. By default,
\code{geobr} stores data files in a temporary directory that exists only
within each R session. If \code{cache = FALSE}, the function will download
the data again and overwrite the local file.}

\item{verbose}{A logical. If \code{TRUE} (the default), the function prints
informative messages and shows download progress bar. If \code{FALSE}, the
function is silent.}
}
\value{
An \verb{"sf" "data.frame"} OR an \code{ArrowObject}
}
\description{
This function downloads either a spatial \code{sf} object with the location of the
municipal seats (sede dos municipios) of state capitals, or a \code{data.frame}
with the names and codes of state capitals. Data downloaded for the latest
available year.
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) withAutoprint(\{ # examplesIf}
# Read spatial data with the  municipal seats of state capitals
capitals_sf <- read_capitals(as_sf = TRUE)

# Read simple data.frame of state capitals
capitals_df <- read_capitals(as_sf = FALSE)
\dontshow{\}) # examplesIf}
}
