% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_capitals.R
\name{read_capitals}
\alias{read_capitals}
\title{Download data of state capitals}
\usage{
read_capitals(as_sf = TRUE, showProgress = TRUE)
}
\arguments{
\item{as_sf}{Logic \code{FALSE} or \code{TRUE}, indicating whether the function should
return a spatial data in \code{sf} format (Defaults to \code{TRUE}) or in a
\code{data.frame} format without spatial information (\code{FALSE}).}

\item{showProgress}{Logical. Defaults to \code{TRUE} display progress bar.}
}
\value{
An \verb{"sf" "data.frame"} object or a \code{"data.frame"}
}
\description{
This function downloads either a spatial \code{sf} object with the location of the
municipal seats (sede dos municipios) of state capitals, or a \code{data.frame}
with the names and codes of state capitals. Data downloaded for the latest
available year.
}
\examples{
\dontrun{ if (interactive()) {
# Read spatial data with the  municipal seats of state capitals
capitals_sf <- read_capitals(as_sf = TRUE)

# Read simple data.frame of state capitals
capitals_df <- read_capitals(as_sf = FALSE)
}}
}
