% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_census_tract.R
\name{read_census_tract}
\alias{read_census_tract}
\title{Download shape files of census sectors of the Brazilian Population Census}
\usage{
read_census_tract(CODE = NULL, year = NULL, zone = "urban")
}
\arguments{
\item{CODE}{One can either pass the 7-digit code of a Municipality or the 2-digit code of a State.If CODE="all", all census sectors of the country are loaded.}

\item{year}{the year of the data download (defaults to 2010)}

\item{zone}{"urban" or "rural" for separation in the year 2000}
}
\description{
Download shape files of census sectors of the Brazilian Population Census
}
\examples{
\dontrun{
# Exemplos
dados <- read_census_tract(year=2010)
dados <- read_census_tract(123,2010)
dados <- read_census_tract("df",2010)
dados <- read_census_tract(1302603,2010)
dados <- read_census_tract(35)
dados <- read_census_tract(14,2010)
dados <- read_census_tract()

# mapa
library(mapview)
mapview(dados)
}




}
\seealso{
Other general area functions: \code{\link{read_meso_region}},
  \code{\link{read_micro_region}},
  \code{\link{read_municipality}},
  \code{\link{read_state}},
  \code{\link{read_statistical_grid}},
  \code{\link{read_weighting_area}}
}
\concept{general area functions}
