% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_census_tract.R
\name{read_census_tract}
\alias{read_census_tract}
\title{Download shape files of census tracts of the Brazilian Population Census}
\usage{
read_census_tract(
  code_tract,
  year = 2010,
  zone = "urban",
  simplified = TRUE,
  showProgress = TRUE
)
}
\arguments{
\item{code_tract}{The 7-digit code of a Municipality. If the two-digit code
or a two-letter uppercase abbreviation of a state is passed, (e.g. 33 or "RJ")
the function will load all census tracts of that state. If \code{code_tract="all"},
all census tracts of the country are loaded.}

\item{year}{Year of the data. Defaults to 2010}

\item{zone}{For census tracts before 2010, 'urban' and 'rural' census tracts are separate data sets.}

\item{simplified}{Logic \code{FALSE} or \code{TRUE}, indicating whether the function
returns the data set with original' resolution or a data set with 'simplified'
borders. Defaults to \code{TRUE}. For spatial analysis and statistics users should
set \code{simplified = FALSE}. Borders have been simplified by removing vertices of
borders using \code{sf::st_simplify()} preserving topology with a \code{dTolerance} of 100.}

\item{showProgress}{Logical. Defaults to \code{TRUE} display progress bar}
}
\description{
Download shape files of census tracts of the Brazilian Population Census
}
\examples{
\dontrun{
# Read rural census tracts for years before 2007
  c <- read_census_tract(code_tract=5201108, year=2000, zone="rural")

# Read all census tracts of a state at a given year
  c <- read_census_tract(code_tract=53, year=2010) # or
  c <- read_census_tract(code_tract="DF", year=2010)
  plot(c)

# Read all census tracts of a municipality at a given year
  c <- read_census_tract(code_tract=5201108, year=2010)
  plot(c)

# Read all census tracts of the country at a given year
  c <- read_census_tract(code_tract="all", year=2010)

}
}
\seealso{
Other general area functions: 
\code{\link{read_amazon}()},
\code{\link{read_biomes}()},
\code{\link{read_conservation_units}()},
\code{\link{read_country}()},
\code{\link{read_health_region}()},
\code{\link{read_immediate_region}()},
\code{\link{read_intermediate_region}()},
\code{\link{read_meso_region}()},
\code{\link{read_micro_region}()},
\code{\link{read_municipality}()},
\code{\link{read_neighborhood}()},
\code{\link{read_region}()},
\code{\link{read_semiarid}()},
\code{\link{read_state}()},
\code{\link{read_statistical_grid}()},
\code{\link{read_weighting_area}()}
}
\concept{general area functions}
