% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_comparable_areas.R
\name{read_comparable_areas}
\alias{read_comparable_areas}
\title{Download spatial data of historically comparable municipalities}
\usage{
read_comparable_areas(
  start_year = 1970,
  end_year = 2010,
  simplified = TRUE,
  showProgress = TRUE
)
}
\arguments{
\item{start_year}{Numeric. Start year to the period.}

\item{end_year}{Numeric. End year to the period. (defaults to 2010)}

\item{simplified}{Logic FALSE or TRUE, indicating whether the function returns the
data set with 'original' resolution or a data set with 'simplified' borders (Defaults to TRUE).
For spatial analysis and statistics users should set simplified = FALSE. Borders have been
simplified by removing vertices of borders using st_simplify{sf} preserving topology with a dTolerance of 100.}

\item{showProgress}{Logical. Defaults to (TRUE) display progress bar}
}
\value{
An \verb{"sf" "data.frame"} object
}
\description{
This function downloads the shape file of minimum comparable area of
municipalities, known in Portuguese as 'Areas minimas comparáveis (AMCs)'.
The data is available for any combination of census years between 1872-2010.
These data sets are generated based on the Stata code originally developed by
Philipp Ehrl \url{https://doi.org/10.1590/0101-416147182phe}, and translated
into \code{R} by the \code{geobr} team. \href{https://doi.org/10.1590/0101-416147182phe}{Philipp Ehrl}
}
\examples{
\donttest{

  amc <- read_comparable_areas(start_year=1970, end_year=2010)
}

}
\seealso{
Other general area functions: 
\code{\link{read_amazon}()},
\code{\link{read_biomes}()},
\code{\link{read_census_tract}()},
\code{\link{read_conservation_units}()},
\code{\link{read_country}()},
\code{\link{read_health_region}()},
\code{\link{read_immediate_region}()},
\code{\link{read_intermediate_region}()},
\code{\link{read_meso_region}()},
\code{\link{read_micro_region}()},
\code{\link{read_municipality}()},
\code{\link{read_neighborhood}()},
\code{\link{read_region}()},
\code{\link{read_semiarid}()},
\code{\link{read_state}()},
\code{\link{read_statistical_grid}()},
\code{\link{read_weighting_area}()}
}
\concept{general area functions}
