% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_conservation_units.R
\name{read_conservation_units}
\alias{read_conservation_units}
\title{Download spatial data of Brazilian environmental conservation units}
\usage{
read_conservation_units(
  date,
  simplified = TRUE,
  as_sf = TRUE,
  showProgress = TRUE,
  cache = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{date}{Numeric. Date of the data in YYYYMM format. It defaults to \code{NULL}
and reads the data from the latest date available.}

\item{simplified}{Logic \code{FALSE} or \code{TRUE}, indicating whether the function
should return the data set with 'original' spatial resolution or a data set
with 'simplified' geometry. Defaults to \code{TRUE}. For spatial analysis and
statistics users should set \code{simplified = FALSE}. Borders have been
simplified by removing vertices of borders using \verb{st_simplify\{sf\}} preserving
topology with a \code{dTolerance} of 100.}

\item{as_sf}{Logic. If \code{TRUE} (the default), the function returns an
\verb{sf data.frame}. If \code{FALSE}, the function returns an arrow dataset.}

\item{showProgress}{Logical. Defaults to \code{TRUE} display progress bar.}

\item{cache}{Logical. Whether the function should read the data cached
locally, which is faster. Defaults to \code{cache = TRUE}. By default,
\code{geobr} stores data files in a temporary directory that exists only
within each R session. If \code{cache = FALSE}, the function will download
the data again and overwrite the local file.}

\item{verbose}{A logical. Whether the function should print informative
messages. Defaults to \code{TRUE}.}
}
\value{
An \verb{"sf" "data.frame"} OR an \code{ArrowObject}
}
\description{
This data set covers the whole of Brazil and it includes the polygons of all
conservation units present in Brazilian territory. The original data and data
dictionary can be found comes from MMA and can be found at "https://dados.mma.gov.br/dataset/unidadesdeconservacao".
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) withAutoprint(\{ # examplesIf}
# Read conservation_units
uc <- read_conservation_units(date = 202503)
\dontshow{\}) # examplesIf}
}
\seealso{
Other general area functions: 
\code{\link{read_census_tract}()}
}
\concept{general area functions}
