% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_disaster_risk_area.R
\name{read_disaster_risk_area}
\alias{read_disaster_risk_area}
\title{Download spatial data of disaster risk areas}
\usage{
read_disaster_risk_area(
  year = NULL,
  code_muni = "all",
  simplified = TRUE,
  as_sf = TRUE,
  showProgress = TRUE,
  cache = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{year}{Numeric. Year of the data in YYYY format. It defaults to \code{NULL}
and reads the data from the latest year available.}

\item{code_muni}{The 7-digit identification code of a municipality. If
\code{code_muni = "all"} (Default), the function downloads all the
disaster risk areas in the country. Alternatively, if a
two-digit identification code or a two-letter uppercase abbreviation
of a state is passed (e.g. \code{33} or \code{"RJ"}), all disaster risk areas of
that state are downloaded. Municipality identification codes can be
consulted with the \code{geobr::lookup_muni()} function.}

\item{simplified}{Logic \code{FALSE} or \code{TRUE}, indicating whether the function
should return the data set with 'original' spatial resolution or a data set
with 'simplified' geometry. Defaults to \code{TRUE}. For spatial analysis and
statistics users should set \code{simplified = FALSE}. Borders have been
simplified by removing vertices of borders using \verb{st_simplify\{sf\}} preserving
topology with a \code{dTolerance} of 100.}

\item{as_sf}{Logic. If \code{TRUE} (the default), the function returns an
\verb{sf data.frame}. If \code{FALSE}, the function returns an arrow dataset.}

\item{showProgress}{Logical. Defaults to \code{TRUE} display progress bar.}

\item{cache}{Logical. Whether the function should read the data cached
locally, which is faster. Defaults to \code{cache = TRUE}. By default,
\code{geobr} stores data files in a temporary directory that exists only
within each R session. If \code{cache = FALSE}, the function will download
the data again and overwrite the local file.}

\item{verbose}{A logical. Whether the function should print informative
messages. Defaults to \code{TRUE}.}
}
\value{
An \verb{"sf" "data.frame"} OR an \code{ArrowObject}
}
\description{
This function reads the the official data of disaster risk areas in Brazil
(currently only available for 2010). It specifically focuses on geodynamic
and hydro-meteorological disasters capable of triggering landslides and floods.
The data set covers the whole country. Each risk area polygon (known as 'BATER')
has unique code id (column 'geo_bater'). The data set brings information on
the extent to which the risk area polygons overlap with census tracts and block
faces (column "acuracia") and number of ris areas within each risk area (column
'num'). Original data were generated by IBGE and CEMADEN. For more information
about the methodology, see deails at \url{https://www.ibge.gov.br/geociencias/organizacao-do-territorio/tipologias-do-territorio/21538-populacao-em-areas-de-risco-no-brasil.html}
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) withAutoprint(\{ # examplesIf}
# Read all disaster risk areas in an specific year
d <- read_disaster_risk_area(year = 2010)

# Read disaster risk areas in a given municipality
d <- read_disaster_risk_area(year = 2010, code_muni = 2927408)

# Read disaster risk areas in a given state
d <- read_disaster_risk_area(year = 2010, code_muni = "AC")
\dontshow{\}) # examplesIf}
}
\seealso{
Other area functions: 
\code{\link{read_amazon}()},
\code{\link{read_biomes}()},
\code{\link{read_capitals}()},
\code{\link{read_comparable_areas}()},
\code{\link{read_country}()},
\code{\link{read_favela}()},
\code{\link{read_health_facilities}()},
\code{\link{read_health_region}()},
\code{\link{read_immediate_region}()},
\code{\link{read_indigenous_land}()},
\code{\link{read_intermediate_region}()},
\code{\link{read_meso_region}()},
\code{\link{read_metro_area}()},
\code{\link{read_micro_region}()},
\code{\link{read_municipality}()},
\code{\link{read_neighborhood}()},
\code{\link{read_pop_arrangements}()},
\code{\link{read_quilombo}()},
\code{\link{read_region}()},
\code{\link{read_schools}()},
\code{\link{read_semiarid}()},
\code{\link{read_state}()},
\code{\link{read_statistical_grid}()},
\code{\link{read_urban_concentrations}()},
\code{\link{read_weighting_area}()}
}
\concept{area functions}
