% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_health_region.R
\name{read_health_region}
\alias{read_health_region}
\title{Download spatial data of Brazilian health regions and health macro regions}
\usage{
read_health_region(
  year,
  code_state = "all",
  group_by = NULL,
  macro = NULL,
  simplified = TRUE,
  as_sf = TRUE,
  showProgress = TRUE,
  cache = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{year}{Numeric. Year of the data in YYYY format. It defaults to \code{NULL}
and reads the data from the latest year available.}

\item{code_state}{The two-digit code of a state or a two-letter uppercase
abbreviation (e.g. 33 or "RJ"). If \code{code_state="all"} (the
default), the function downloads all states.}

\item{group_by}{String. When \code{group_by = NULL} (Default), the function returns
the geometries of municipalities. When \code{group_by = "micro"}, the output
geometries are aggregated to return polygons of micro health regions.
Alternatively, \code{group_by = "macro"} returns polygons of macro health
regions.}

\item{macro}{The argument \code{macro} has been deprecated.}

\item{simplified}{Logic \code{FALSE} or \code{TRUE}, indicating whether the function
should return the data set with 'original' spatial resolution or a data set
with 'simplified' geometry. Defaults to \code{TRUE}. For spatial analysis and
statistics users should set \code{simplified = FALSE}. Borders have been
simplified by removing vertices of borders using \verb{st_simplify\{sf\}} preserving
topology with a \code{dTolerance} of 100.}

\item{as_sf}{Logic. If \code{TRUE} (the default), the function returns an
\verb{sf data.frame}. If \code{FALSE}, the function returns an arrow dataset.}

\item{showProgress}{Logical. Defaults to \code{TRUE} display progress bar.}

\item{cache}{Logical. Whether the function should read the data cached
locally, which is faster. Defaults to \code{cache = TRUE}. By default,
\code{geobr} stores data files in a temporary directory that exists only
within each R session. If \code{cache = FALSE}, the function will download
the data again and overwrite the local file.}

\item{verbose}{A logical. Whether the function should print informative
messages. Defaults to \code{TRUE}.}
}
\value{
An \verb{"sf" "data.frame"} OR an \code{ArrowObject}
}
\description{
Health regions are used to guide the the regional and state planning of health
services. Macro health regions, in particular, are used to guide the planning
of high complexity #' health services. These services involve larger economics
of scale and are concentrated in few municipalities because they are generally
more technology intensive, costly and face shortages of specialized
professionals. A macro region comprises one or more health regions.
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) withAutoprint(\{ # examplesIf}
# Read health regions for a given year
health_muni <- read_health_region(year = 2024)

# Read health regions with the geometries of micro regions
health_micro <- read_health_region(
  year = 2024,
  group_by = "micro"
  )

# Read health regions with the geometries of micro regions
health_macro <- read_health_region(
  year = 2024,
  group_by = "macro"
)
\dontshow{\}) # examplesIf}
}
\seealso{
Other area functions: 
\code{\link{read_amazon}()},
\code{\link{read_biomes}()},
\code{\link{read_capitals}()},
\code{\link{read_comparable_areas}()},
\code{\link{read_country}()},
\code{\link{read_disaster_risk_area}()},
\code{\link{read_favela}()},
\code{\link{read_health_facilities}()},
\code{\link{read_immediate_region}()},
\code{\link{read_indigenous_land}()},
\code{\link{read_intermediate_region}()},
\code{\link{read_meso_region}()},
\code{\link{read_metro_area}()},
\code{\link{read_micro_region}()},
\code{\link{read_municipality}()},
\code{\link{read_neighborhood}()},
\code{\link{read_pop_arrangements}()},
\code{\link{read_quilombo}()},
\code{\link{read_region}()},
\code{\link{read_schools}()},
\code{\link{read_semiarid}()},
\code{\link{read_state}()},
\code{\link{read_statistical_grid}()},
\code{\link{read_urban_concentrations}()},
\code{\link{read_weighting_area}()}
}
\concept{area functions}
