% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_immediate_region.R
\name{read_immediate_region}
\alias{read_immediate_region}
\title{Download spatial data of Brazil's Immediate Geographic Areas}
\usage{
read_immediate_region(
  year,
  code_immediate = "all",
  simplified = TRUE,
  output = "sf",
  showProgress = TRUE,
  cache = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{year}{Numeric. Year of the data in \code{YYYY} format.}

\item{code_immediate}{6-digit code of an immediate region. If the two-digit
code or a two-letter uppercase abbreviation of a state is passed, (e.g.
33 or "RJ") the function will load all immediate regions of that state.
If \code{code_immediate="all"} (Default), the function downloads all
immediate regions of the country.}

\item{simplified}{Logic \code{FALSE} or \code{TRUE}, indicating whether the function
should return the data set with 'original' spatial resolution or a data set
with 'simplified' geometry. Defaults to \code{TRUE}. For spatial analysis and
statistics users should set \code{simplified = FALSE}. Borders have been
simplified by removing vertices of borders using \verb{st_simplify\{sf\}} preserving
topology with a \code{dTolerance} of 100.}

\item{output}{String. Type of object returned by the function. Defaults to
\code{"sf"}, which loads the data into memory as an sf object. Alternatively,
\code{"duckdb"} returns a lazy spatial table backed by DuckDB via the
duckspatial package, and \code{"arrow"} returns an Arrow dataset. Both
\code{"duckdb"} and \code{"arrow"} support out-of-memory processing of large
data sets.}

\item{showProgress}{Logical. Defaults to \code{TRUE} display progress bar.}

\item{cache}{Logical. Whether the function should read the data cached
locally, which is faster. Defaults to \code{cache = TRUE}. By default,
\code{geobr} stores data files in a temporary directory that exists only
within each R session. If \code{cache = FALSE}, the function will download
the data again and overwrite the local file.}

\item{verbose}{A logical. If \code{TRUE} (the default), the function prints
informative messages and shows download progress bar. If \code{FALSE}, the
function is silent.}
}
\value{
An \verb{"sf" "data.frame"} OR an \code{ArrowObject}
}
\description{
The Immediate Geographic Areas are part of the geographic division of
Brazil created after 2017 by IBGE. These regions were created to replace the
"Micro Regions" division. Data at scale 1:250,000.
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) withAutoprint(\{ # examplesIf}
# Read an specific immediate region
im <- read_immediate_region(code_immediate = 110006, year = 2024)

# Read immediate regions of a state
im <- read_immediate_region(code_immediate = "AM", year = 2024)
im <- read_immediate_region(code_immediate = 12, year = 2024)

# Read all immediate regions of the country
im <- read_immediate_region(code_immediate = "all", year = 2024)
\dontshow{\}) # examplesIf}
}
