% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_intermediate_region.R
\name{read_intermediate_region}
\alias{read_intermediate_region}
\title{Download spatial data of Brazil's Intermediate Geographic Areas}
\usage{
read_intermediate_region(
  code_intermediate = "all",
  year = 2019,
  simplified = TRUE,
  showProgress = TRUE
)
}
\arguments{
\item{code_intermediate}{4-digit code of an intermediate region. If the
two-digit code or a two-letter uppercase abbreviation of a state is passed,
(e.g. 33 or "RJ") the function will load all intermediate regions of that
state. If \code{code_intermediate="all"} (Default), all intermediate regions of
the country are loaded.}

\item{year}{A year number in YYYY format. Defaults to \code{2019}}

\item{simplified}{Logic \code{FALSE} or \code{TRUE}, indicating whether the function
returns the data set with original' resolution or a data set with 'simplified'
borders. Defaults to \code{TRUE}. For spatial analysis and statistics users should
set \code{simplified = FALSE}. Borders have been simplified by removing vertices of
borders using \code{sf::st_simplify()} preserving topology with a \code{dTolerance} of 100.}

\item{showProgress}{Logical. Defaults to \code{TRUE} display progress bar}
}
\value{
An \verb{"sf" "data.frame"} object
}
\description{
The intermediate Geographic Areas are part of the geographic division of
Brazil created in 2017 by IBGE. These regions were created to replace the
"Meso Regions" division. Data at scale 1:250,000, using Geodetic reference
system "SIRGAS2000" and CRS(4674)
}
\examples{
\dontrun{
# Read an specific intermediate region
  im <- read_intermediate_region(code_intermediate=1202)

# Read intermediate regions of a state
  im <- read_intermediate_region(code_intermediate=12)
  im <- read_intermediate_region(code_intermediate="AM")

# Read all intermediate regions of the country
  im <- read_intermediate_region()
  im <- read_intermediate_region(code_intermediate="all")
}
}
\seealso{
Other general area functions: 
\code{\link{read_amazon}()},
\code{\link{read_biomes}()},
\code{\link{read_census_tract}()},
\code{\link{read_conservation_units}()},
\code{\link{read_country}()},
\code{\link{read_health_region}()},
\code{\link{read_immediate_region}()},
\code{\link{read_meso_region}()},
\code{\link{read_micro_region}()},
\code{\link{read_municipality}()},
\code{\link{read_neighborhood}()},
\code{\link{read_region}()},
\code{\link{read_semiarid}()},
\code{\link{read_state}()},
\code{\link{read_statistical_grid}()},
\code{\link{read_weighting_area}()}
}
\concept{general area functions}
