% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_meso_region.R
\name{read_meso_region}
\alias{read_meso_region}
\title{Download shape files of meso region.}
\usage{
read_meso_region(code_meso, year = NULL)
}
\arguments{
\item{code_meso}{The 4-digit code of a meso region. If a two-digit code of a state is passed,
the function will load all meso regions of that state. If code_meso="all", all meso regions of the country are loaded.}

\item{year}{Year of the data (defaults to 2010)}
}
\description{
Download shape files of meso region.
}
\examples{
\dontrun{

library(geobr)

# Read all meso regions of a state at a given year
  meso <- read_meso_region(code_meso=12, year=2017)

# Read all meso regions of the country at a given year
  meso <- read_meso_region(code_meso="all", year=2010)
}

}
\seealso{
Other general area functions: \code{\link{read_census_tract}},
  \code{\link{read_micro_region}},
  \code{\link{read_municipality}},
  \code{\link{read_state}},
  \code{\link{read_statistical_grid}},
  \code{\link{read_weighting_area}}
}
\concept{general area functions}
