% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_micro_region.R
\name{read_micro_region}
\alias{read_micro_region}
\title{Download spatial data of micro regions}
\usage{
read_micro_region(
  code_micro = "all",
  year = 2010,
  simplified = TRUE,
  showProgress = TRUE
)
}
\arguments{
\item{code_micro}{5-digit code of a micro region. If the two-digit code or a
two-letter uppercase abbreviation of a state is passed, (e.g. 33 or
"RJ") the function will load all micro regions of that state. If
\code{code_micro="all"} (Default), the function downloads all micro regions of the
country.}

\item{year}{Numeric. Year of the data in YYYY format. Defaults to \code{2010}.}

\item{simplified}{Logic \code{FALSE} or \code{TRUE}, indicating whether the function
should return the data set with 'original' spatial resolution or a data set
with 'simplified' geometry. Defaults to \code{TRUE}. For spatial analysis and
statistics users should set \code{simplified = FALSE}. Borders have been
simplified by removing vertices of borders using \verb{st_simplify\{sf\}} preserving
topology with a \code{dTolerance} of 100.}

\item{showProgress}{Logical. Defaults to \code{TRUE} display progress bar.}
}
\value{
An \verb{"sf" "data.frame"} object
}
\description{
Data at scale 1:250,000, using Geodetic reference system "SIRGAS2000" and CRS(4674)
}
\examples{
\dontrun{ if (interactive()) {
# Read an specific micro region a given year
  micro <- read_micro_region(code_micro=11008, year=2018)

# Read micro regions of a state at a given year
  micro <- read_micro_region(code_micro=12, year=2017)
  micro <- read_micro_region(code_micro="AM", year=2000)

# Read all micro regions at a given year
  micro <- read_micro_region(code_micro="all", year=2010)
}}
}
\seealso{
Other general area functions: 
\code{\link{read_amazon}()},
\code{\link{read_biomes}()},
\code{\link{read_census_tract}()},
\code{\link{read_comparable_areas}()},
\code{\link{read_conservation_units}()},
\code{\link{read_country}()},
\code{\link{read_health_region}()},
\code{\link{read_immediate_region}()},
\code{\link{read_intermediate_region}()},
\code{\link{read_meso_region}()},
\code{\link{read_municipality}()},
\code{\link{read_neighborhood}()},
\code{\link{read_region}()},
\code{\link{read_semiarid}()},
\code{\link{read_state}()},
\code{\link{read_statistical_grid}()},
\code{\link{read_weighting_area}()}
}
\concept{general area functions}
