% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_micro_region.R
\name{read_micro_region}
\alias{read_micro_region}
\title{Download shape files of micro region.}
\usage{
read_micro_region(cod_micro, year = NULL)
}
\arguments{
\item{cod_micro}{5-digit code of a micro region. If the two-digit code of a state is passed,
the function will load all micro regions of that state. If cod_micro="all", all micro regions of the country are loaded.}

\item{year}{Year of the data (defaults to 2010)}
}
\description{
Download shape files of micro region.
}
\examples{
\dontrun{

library(geobr)

# Read all micro regions of a state at a given year
  micro <- read_micro_region(cod_micro=12, year=2017)

# Read all micro regions at a given year
  micro <- read_micro_region(cod_micro="all", year=2010)
}


}
\seealso{
Other general area functions: \code{\link{read_census_tract}},
  \code{\link{read_meso_region}},
  \code{\link{read_municipality}},
  \code{\link{read_state}},
  \code{\link{read_statistical_grid}},
  \code{\link{read_weighting_area}}
}
\concept{general area functions}
