% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_municipal_seat.R
\name{read_municipal_seat}
\alias{read_municipal_seat}
\title{Download spatial data of municipal seats (sede dos municipios) of Brazil}
\usage{
read_municipal_seat(
  year,
  code_muni = "all",
  as_sf = TRUE,
  showProgress = TRUE,
  cache = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{year}{Numeric. Year of the data in YYYY format. It defaults to \code{NULL}
and reads the data from the latest year available.}

\item{code_muni}{The 7-digit code of a municipality. If \code{code_muni = "all"}
(Default), the function downloads all the data available in the
country. Alternatively, if a two-digit state code or a two-letter
uppercase abbreviation of a state is passed (e.g. \code{33} or \code{"RJ"}), all
data of that state are downloaded. Municipality codes can be consulted
with the \code{geobr::lookup_muni()} function.}

\item{as_sf}{Logic. If \code{TRUE} (the default), the function returns an
\verb{sf data.frame}. If \code{FALSE}, the function returns an arrow dataset.}

\item{showProgress}{Logical. Defaults to \code{TRUE} display progress bar.}

\item{cache}{Logical. Whether the function should read the data cached
locally, which is faster. Defaults to \code{cache = TRUE}. By default,
\code{geobr} stores data files in a temporary directory that exists only
within each R session. If \code{cache = FALSE}, the function will download
the data again and overwrite the local file.}

\item{verbose}{A logical. Whether the function should print informative
messages. Defaults to \code{TRUE}.}
}
\value{
An \verb{"sf" "data.frame"} OR an \code{ArrowObject}
}
\description{
This function reads the official data on the municipal seats (sede dos
municipios) of Brazil. The data brings the geographical coordinates (lat lon)
of municipal seats for various years since 1872. Original data by the
Brazilian Institute of Geography and Statistics (IBGE).
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) withAutoprint(\{ # examplesIf}
# Read municipal seats in an specific year
m <- read_municipal_seat(year = 2022)
\dontshow{\}) # examplesIf}
}
