% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_neighborhood.R
\name{read_neighborhood}
\alias{read_neighborhood}
\title{Download spatial data of neighborhood limits of Brazilian municipalities}
\usage{
read_neighborhood(year = 2010, simplified = TRUE, showProgress = TRUE)
}
\arguments{
\item{year}{Year of the data. Defaults to \code{2010}}

\item{simplified}{Logic \code{FALSE} or \code{TRUE}, indicating whether the function
returns the data set with 'original' resolution or a data set with 'simplified'
borders. Defaults to \code{TRUE}. For spatial analysis and statistics users should
set \code{simplified = FALSE}. Borders have been simplified by removing vertices of
borders using \verb{st_simplify\{sf\}} preserving topology with a \code{dTolerance} of 100.}

\item{showProgress}{Logical. Defaults to `TRUE`` display progress bar}
}
\value{
An \verb{"sf" "data.frame"} object
}
\description{
This data set includes the neighborhood limits of 720 Brazilian municipalities.
It is based on aggregations of the census tracts from the Brazilian
census. Only 2010 data is currently available.
}
\examples{
\dontrun{ if (interactive()) {
# Read neighborhoods of Brazilian municipalities
n <- read_neighborhood(year=2010)
}}
}
\seealso{
Other general area functions: 
\code{\link{read_amazon}()},
\code{\link{read_biomes}()},
\code{\link{read_census_tract}()},
\code{\link{read_comparable_areas}()},
\code{\link{read_conservation_units}()},
\code{\link{read_country}()},
\code{\link{read_health_region}()},
\code{\link{read_immediate_region}()},
\code{\link{read_intermediate_region}()},
\code{\link{read_meso_region}()},
\code{\link{read_micro_region}()},
\code{\link{read_municipality}()},
\code{\link{read_region}()},
\code{\link{read_semiarid}()},
\code{\link{read_state}()},
\code{\link{read_statistical_grid}()},
\code{\link{read_weighting_area}()}
}
\concept{general area functions}
