% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pop_arrangements.R
\name{read_pop_arrangements}
\alias{read_pop_arrangements}
\title{Download population arrangements in Brazil}
\usage{
read_pop_arrangements(
  year,
  code_state = "all",
  simplified = TRUE,
  output = "sf",
  showProgress = TRUE,
  cache = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{year}{Numeric. Year of the data in YYYY format. It defaults to \code{NULL}
and reads the data from the latest year available.}

\item{code_state}{The two-digit code of a state or a two-letter uppercase
abbreviation (e.g. 33 or "RJ"). If \code{code_state="all"} (the
default), the function downloads all states.}

\item{simplified}{Logic \code{FALSE} or \code{TRUE}, indicating whether the function
should return the data set with 'original' spatial resolution or a data set
with 'simplified' geometry. Defaults to \code{TRUE}. For spatial analysis and
statistics users should set \code{simplified = FALSE}. Borders have been
simplified by removing vertices of borders using \verb{st_simplify\{sf\}} preserving
topology with a \code{dTolerance} of 100.}

\item{output}{String. Whether the output should be an \code{"sf"} object loaded to
memory (the default), or an \code{"arrow"} arrow data set.}

\item{showProgress}{Logical. Defaults to \code{TRUE} display progress bar.}

\item{cache}{Logical. Whether the function should read the data cached
locally, which is faster. Defaults to \code{cache = TRUE}. By default,
\code{geobr} stores data files in a temporary directory that exists only
within each R session. If \code{cache = FALSE}, the function will download
the data again and overwrite the local file.}

\item{verbose}{A logical. If \code{TRUE} (the default), the function prints
informative messages and shows download progress bar. If \code{FALSE}, the
function is silent.}
}
\value{
An \verb{"sf" "data.frame"} OR an \code{ArrowObject}
}
\description{
This function reads the official data on the population arrangements (Arranjos
Populacionais) in Brazil. Original data by the Brazilian Institute of Geography
and Statistics (IBGE). More information about the methodology at \url{https://www.ibge.gov.br/geociencias/organizacao-do-territorio/divisao-regional/15782-arranjos-populacionais-e-concentracoes-urbanas-do-brasil.html}
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) withAutoprint(\{ # examplesIf}
# Read population arrangements in an specific year
pa <- read_pop_arrangements(year = 2010)
\dontshow{\}) # examplesIf}
}
