% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pop_arrangements.R
\name{read_pop_arrangements}
\alias{read_pop_arrangements}
\title{Download population arrangements in Brazil}
\usage{
read_pop_arrangements(
  year = NULL,
  simplified = TRUE,
  showProgress = TRUE,
  cache = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{year}{Numeric. Year of the data in YYYY format. It defaults to \code{NULL}
and reads the data from the latest year available.}

\item{simplified}{Logic \code{FALSE} or \code{TRUE}, indicating whether the function
should return the data set with 'original' spatial resolution or a data set
with 'simplified' geometry. Defaults to \code{TRUE}. For spatial analysis and
statistics users should set \code{simplified = FALSE}. Borders have been
simplified by removing vertices of borders using \verb{st_simplify\{sf\}} preserving
topology with a \code{dTolerance} of 100.}

\item{showProgress}{Logical. Defaults to \code{TRUE} display progress bar.}

\item{cache}{Logical. Whether the function should read the data cached
locally, which is faster. Defaults to \code{cache = TRUE}. By default,
\code{geobr} stores data files in a temporary directory that exists only
within each R session. If \code{cache = FALSE}, the function will download
the data again and overwrite the local file.}

\item{verbose}{A logical. Whether the function should print informative
messages. Defaults to \code{TRUE}.}
}
\value{
An \verb{"sf" "data.frame"} OR an \code{ArrowObject}
}
\description{
This function reads the official data on population arrangements (Arranjos
Populacionais) of Brazil. Original data were generated by the Institute of
Geography and Statistics (IBGE)  For more information about the methodology,
see details at \url{https://www.ibge.gov.br/apps/arranjos_populacionais/2015/pdf/publicacao.pdf}
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) withAutoprint(\{ # examplesIf}
# Read urban footprint of Brazilian cities in an specific year
uc <- read_pop_arrangements(year=2015)
\dontshow{\}) # examplesIf}
}
