% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_region.R
\name{read_region}
\alias{read_region}
\title{Download shape file of Brazil Regions as sf objects.}
\usage{
read_region(year = NULL)
}
\arguments{
\item{year}{Year of the data (defaults to 2010)}
}
\description{
Data at scale 1:250,000, using Geodetic reference system "SIRGAS2000" and CRS(4674)
}
\examples{
\donttest{

library(geobr)

# Read specific year
  reg <- read_region(year=2018)

}
}
\seealso{
Other general area functions: \code{\link{read_amazon}},
  \code{\link{read_biomes}},
  \code{\link{read_census_tract}},
  \code{\link{read_conservation_units}},
  \code{\link{read_country}},
  \code{\link{read_immediate_region}},
  \code{\link{read_intermediate_region}},
  \code{\link{read_meso_region}},
  \code{\link{read_micro_region}},
  \code{\link{read_municipality}},
  \code{\link{read_semiarid}}, \code{\link{read_state}},
  \code{\link{read_statistical_grid}},
  \code{\link{read_weighting_area}}
}
\concept{general area functions}
