% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_schools.R
\name{read_schools}
\alias{read_schools}
\title{Download geolocated data of schools}
\usage{
read_schools(
  year,
  code_muni = "all",
  as_sf = TRUE,
  showProgress = TRUE,
  cache = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{year}{Numeric. Year of the data in YYYY format. It defaults to \code{NULL}
and reads the data from the latest year available.}

\item{code_muni}{The 7-digit code of a municipality. If \code{code_muni = "all"}
(Default), the function downloads all the data available in the
country. Alternatively, if a two-digit state code or a two-letter
uppercase abbreviation of a state is passed (e.g. \code{33} or \code{"RJ"}), all
data of that state are downloaded. Municipality codes can be consulted
with the \code{geobr::lookup_muni()} function.}

\item{as_sf}{Logic. If \code{TRUE} (the default), the function returns an
\verb{sf data.frame}. If \code{FALSE}, the function returns an arrow dataset.}

\item{showProgress}{Logical. Defaults to \code{TRUE} display progress bar.}

\item{cache}{Logical. Whether the function should read the data cached
locally, which is faster. Defaults to \code{cache = TRUE}. By default,
\code{geobr} stores data files in a temporary directory that exists only
within each R session. If \code{cache = FALSE}, the function will download
the data again and overwrite the local file.}

\item{verbose}{A logical. Whether the function should print informative
messages. Defaults to \code{TRUE}.}
}
\value{
An \verb{"sf" "data.frame"} OR an \code{ArrowObject}
}
\description{
Data comes from the School Census collected by INEP, the National Institute
for Educational Studies and Research Anisio Teixeira. The date of the last
data update is registered in the database in the column 'date_update'. These
data uses Geodetic reference system "SIRGAS2000" and CRS(4674). The coordinates
of each school if collected by INEP. Periodically the coordinates are revised
with the objective of improving the quality of the data. More information
available at \url{https://www.gov.br/inep/pt-br/acesso-a-informacao/dados-abertos/inep-data/catalogo-de-escolas/}
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) withAutoprint(\{ # examplesIf}
# Read all schools in the country
s <- read_schools( year = 2020)
\dontshow{\}) # examplesIf}
}
