% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_state.R
\name{read_state}
\alias{read_state}
\title{Download shape files of Brazilian states}
\usage{
read_state(cod_uf, year = NULL)
}
\arguments{
\item{cod_uf}{The two-digit code of a state. If cod_uf="all", all states will be loaded.}

\item{year}{Year of the data (defaults to 2010)}
}
\description{
Download shape files of Brazilian states
}
\examples{
\dontrun{

library(geobr)

# Read specific municipality at a given year
  uf <- read_state(cod_uf=12, year=2017)

# Read all states at a given year
  ufs <- read_state(cod_uf="all", year=2010)

}
}
\seealso{
Other general area functions: \code{\link{read_census_tract}},
  \code{\link{read_meso_region}},
  \code{\link{read_micro_region}},
  \code{\link{read_municipality2}},
  \code{\link{read_municipality}},
  \code{\link{read_statistical_grid}},
  \code{\link{read_weighting_area}}
}
