% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_statistical_grid.R
\name{read_statistical_grid}
\alias{read_statistical_grid}
\title{Download shape files of IBGE's statistical grid (200 x 200 meters)}
\usage{
read_statistical_grid(year = NULL, cod_grid = NULL)
}
\arguments{
\item{year}{Year of the data (defaults to 2010). The only year available thus far is 2010.}

\item{cod_grid}{The 7-digit code of a grid quadrant If the two-letter abbreviation of a state is used,
the function will load all grid gradrants that intersect with that state. If cod_grid="all", the grid of the whole country will be loaded.}
}
\description{
Download shape files of IBGE's statistical grid (200 x 200 meters)
}
\examples{
\dontrun{

library(geobr)

# Read specific municipality at a given year
  grid <- read_statistical_grid(cod_grid = 45, year=2010)

# Read all municipalities of a state at a given year
  state_grid <- read_statistical_grid(cod_grid = "RJ")

}
}
\seealso{
Other general area functions: \code{\link{read_census_tract}},
  \code{\link{read_meso_region}},
  \code{\link{read_micro_region}},
  \code{\link{read_municipality}},
  \code{\link{read_state}},
  \code{\link{read_weighting_area}}
}
\concept{general area functions}
