% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_uf.R
\name{read_uf}
\alias{read_uf}
\title{Download shape files of Brazilian states}
\usage{
read_uf(year = NULL, cod_uf = NULL)
}
\arguments{
\item{year}{the year of the data download (defaults to 2010)}

\item{cod_uf}{2-digit code of a state. If cod_uf="all", all states will be loaded.}
}
\description{
Download shape files of Brazilian states
}
\examples{
\dontrun{

library(geobr)

# Read specific municipality at a given year
  mun <- read_uf(cod_uf=12, year=2017)

# Read all states at a given year
  mun <- read_uf(cod_uf="all", year=2010)

}
}
\seealso{
Other general area functions: \code{\link{read_areaponderacao}},
  \code{\link{read_mesorregiao}},
  \code{\link{read_microregiao}},
  \code{\link{read_municipio}},
  \code{\link{read_setorcensitario}}
}
\concept{general area functions}
