% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_urban_area.R
\name{read_urban_area}
\alias{read_urban_area}
\title{Download spatial data of urbanized areas in Brazil}
\usage{
read_urban_area(
  year = NULL,
  code_muni = "all",
  simplified = TRUE,
  as_sf = TRUE,
  showProgress = TRUE,
  cache = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{year}{Numeric. Year of the data in YYYY format. It defaults to \code{NULL}
and reads the data from the latest year available.}

\item{code_muni}{The 7-digit identification code of a municipality. If
\code{code_muni = "all"} (Default), the function downloads all the urban
footprints data available in the country. Alternatively, if a
two-digit identification code or a two-letter uppercase abbreviation of
a state is passed (e.g. \code{33} or \code{"RJ"}), all urban footprints of that
state are downloaded. Municipality identification codes can be consulted
with the \code{geobr::lookup_muni()} function.}

\item{simplified}{Logic \code{FALSE} or \code{TRUE}, indicating whether the function
should return the data set with 'original' spatial resolution or a data set
with 'simplified' geometry. Defaults to \code{TRUE}. For spatial analysis and
statistics users should set \code{simplified = FALSE}. Borders have been
simplified by removing vertices of borders using \verb{st_simplify\{sf\}} preserving
topology with a \code{dTolerance} of 100.}

\item{as_sf}{Logic. If \code{TRUE} (the default), the function returns an
\verb{sf data.frame}. If \code{FALSE}, the function returns an arrow dataset.}

\item{showProgress}{Logical. Defaults to \code{TRUE} display progress bar.}

\item{cache}{Logical. Whether the function should read the data cached
locally, which is faster. Defaults to \code{cache = TRUE}. By default,
\code{geobr} stores data files in a temporary directory that exists only
within each R session. If \code{cache = FALSE}, the function will download
the data again and overwrite the local file.}

\item{verbose}{A logical. Whether the function should print informative
messages. Defaults to \code{TRUE}.}
}
\value{
An \verb{"sf" "data.frame"} OR an \code{ArrowObject}
}
\description{
This function reads the official data on the urban footprint of Brazilian
cities. Original data by the Brazilian Institute of Geography and Statistics
(IBGE)  For more information about the methodology, see details
at \url{https://biblioteca.ibge.gov.br/visualizacao/livros/liv100639.pdf}
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) withAutoprint(\{ # examplesIf}
# Read urban footprint of Brazilian cities in an specific year
d <- read_urban_area(year = 2015)
\dontshow{\}) # examplesIf}
}
