% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_weighting_area.R
\name{read_weighting_area}
\alias{read_weighting_area}
\title{Download shape files of Census Weighting Areas (área de ponderação) of the Brazilian Population Census}
\usage{
read_weighting_area(CODE = NULL, year = NULL)
}
\arguments{
\item{CODE}{One can either pass the 7-digit code of a Municipality or the 2-digit code of a State. The function will load the shape files of all weighting areas in the specified geography}

\item{year}{the year of the data download (defaults to 2010)}
}
\description{
Download shape files of Census Weighting Areas (área de ponderação) of the Brazilian Population Census
}
\examples{
\dontrun{

library(geobr)

dados <- read_weighting_area(year=2010)
dados <- read_weighting_area(3500000,2010)
dados <- read_weighting_area(123,2010)
dados <- read_weighting_area("df",2010)
dados <- read_weighting_area(1302603,2010)
dados <- read_weighting_area(35)
dados <- read_weighting_area(14,2010)
dados <- read_weighting_area()

# map it
library(mapview)
mapview(dados)
}




}
\seealso{
Other general area functions: \code{\link{read_census_tract}},
  \code{\link{read_meso_region}},
  \code{\link{read_micro_region}},
  \code{\link{read_municipality}},
  \code{\link{read_state}},
  \code{\link{read_statistical_grid}}
}
\concept{general area functions}
