% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_data_type}
\alias{select_data_type}
\title{Support functions for geobr
Select data type: 'original' or 'simplified' (default)}
\usage{
select_data_type(temp_meta, tp)
}
\description{
Support functions for geobr
Select data type: 'original' or 'simplified' (default)
}
