% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_data_type}
\alias{select_data_type}
\title{Select data type: 'original' or 'simplified' (default)}
\usage{
select_data_type(temp_meta, simplified = NULL)
}
\arguments{
\item{temp_meta}{A dataframe with the file_url addresses of geobr datasets}

\item{simplified}{Logical TRUE or FALSE indicating  whether the function returns the 'original' dataset with high resolution or a dataset with 'simplified' borders (Defaults to TRUE)}
}
\description{
Select data type: 'original' or 'simplified' (default)
}
\keyword{internal}
