% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_data_type}
\alias{select_data_type}
\title{Select data type: 'original' or 'simplified' (default)}
\usage{
select_data_type(temp_meta, tp)
}
\arguments{
\item{temp_meta}{A dataframe with the url addresses of geobr datasets}

\item{tp}{A string indicating whether the function returns the 'original' dataset with high resolution or a dataset with 'simplified' borders (Default)}
}
\description{
Select data type: 'original' or 'simplified' (default)
}
\seealso{
Other support functions: 
\code{\link{download_gpkg}()}
}
\concept{support functions}
