% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_metadata}
\alias{select_metadata}
\title{Select metadata}
\usage{
select_metadata(geography, year = NULL, simplified = NULL)
}
\arguments{
\item{geography}{Which geography will be downloaded.}

\item{year}{Year of the dataset (passed by read_ function).}

\item{simplified}{Logical TRUE or FALSE indicating  whether the function
returns the 'original' dataset with high resolution or a dataset with
'simplified' borders (Defaults to TRUE).}
}
\description{
Select metadata
}
\examples{
\dontrun{ if (interactive()) {

library(geobr)

df <- download_metadata()

}}

}
\keyword{internal}
