% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_metadata}
\alias{select_metadata}
\title{Select metadata}
\usage{
select_metadata(geography, year = NULL, simplified = NULL)
}
\arguments{
\item{geography}{Which geography will be downloaded}

\item{year}{Year of the dataset (passed by red_ function)}

\item{simplified}{Logical TRUE or FALSE indicating  whether the function returns the 'original' dataset with high resolution or a dataset with 'simplified' borders (Defaults to TRUE)}
}
\description{
Select metadata
}
\examples{
\donttest{

library(geobr)

df <- download_metadata()

}

}
\seealso{
Other support functions: 
\code{\link{check_connection}()},
\code{\link{download_gpkg}()},
\code{\link{load_gpkg}()},
\code{\link{select_data_type}()},
\code{\link{select_year_input}()}
}
\concept{support functions}
