% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_year_input}
\alias{select_year_input}
\title{Select year input}
\usage{
select_year_input(temp_meta, y = year)
}
\arguments{
\item{temp_meta}{A dataframe with the file_url addresses of geobr datasets}

\item{y}{Year of the dataset (passed by red_ function)}
}
\description{
Select year input
}
\seealso{
Other support functions: 
\code{\link{check_connection}()},
\code{\link{download_gpkg}()},
\code{\link{load_gpkg}()},
\code{\link{select_data_type}()},
\code{\link{select_metadata}()}
}
\concept{support functions}
